+++ wirecardpaymentgateway/models/Wirecard/PaymentGateway/Payment.php
@@ -31,8 +31,6 @@
 
 require_once __DIR__.'/../../../vendor/autoload.php';
 require_once __DIR__.'/../../../libraries/Logger.php';
-require_once __DIR__.'/../../../libraries/ConfigurationSettings.php';
-
 require_once __DIR__.'/Cart.php';
 
 use Wirecard\PaymentSdk\Config\Config;
@@ -60,7 +58,7 @@ class WirecardPaymentGatewayPayment
 
     public function isAvailable()
     {
-        return (bool)ConfigurationSettings::getConfigValue($this->paymentMethod, 'enable_method');
+        return (bool)$this->module->getConfigValue($this->paymentMethod, 'enable_method');
     }
 
     public function getName()
@@ -88,11 +86,11 @@ class WirecardPaymentGatewayPayment
         $currency = new CurrencyCore($this->module->getContext()->cart->id_currency);
         $currencyIsoCode = $currency->iso_code;
 
-        $baseUrl = ConfigurationSettings::getConfigValue($this->paymentMethod, 'wirecard_server_url');
-        $httpUser = ConfigurationSettings::getConfigValue($this->paymentMethod, 'http_user');
-        $httpPass = ConfigurationSettings::getConfigValue($this->paymentMethod, 'http_password');
-        $MAID = ConfigurationSettings::getConfigValue($this->paymentMethod, 'maid');
-        $key = ConfigurationSettings::getConfigValue($this->paymentMethod, 'secret') ;
+        $baseUrl = $this->module->getConfigValue($this->paymentMethod, 'wirecard_server_url');
+        $httpUser = $this->module->getConfigValue($this->paymentMethod, 'http_user');
+        $httpPass = $this->module->getConfigValue($this->paymentMethod, 'http_password');
+        $MAID = $this->module->getConfigValue($this->paymentMethod, 'maid');
+        $key = $this->module->getConfigValue($this->paymentMethod, 'secret') ;
 
         $this->connection = new Config($baseUrl, $httpUser, $httpPass, $currencyIsoCode);
